requirements/local.txt
-r ./base.txt

psycopg2-binary==2.8.6

# Testing
pytest==5.3.2
pytest-cov==2.8.1
pytest-django==3.8.0
mixer==6.1.3

# Linting
mypy==0.761
flake8==3.7.9
black==19.10b0
pylint-django==2.0.13

# Django
django-debug-toolbar==2.1
django-extensions==2.2.5

# Utils
python-dotenv==0.12.0
